requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
polars==0.19.0
python-multipart==0.0.6
python-dateutil==2.8.2
